ProcessNoiseSystemMap
Introduces a new ProcessNoiseSystemMap which calculates expectations with respect to process noise using the Koopman approach. Right now it always uses the Kosambi-Karhunen-Loevre expansion as the representation of the noise process into discrete random variables, but this can be expanded for other forms as well.
