+++ README.md
@@ -2,8 +2,8 @@
 
 [![Join the chat at https://gitter.im/JuliaDiffEq/Lobby](https://badges.gitter.im/JuliaDiffEq/Lobby.svg)](https://gitter.im/JuliaDiffEq/Lobby?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge)
 [![Build Status](https://github.com/SciML/SciMLExpectations.jl/workflows/CI/badge.svg)](https://github.com/SciML/SciMLExpectations.jl/actions?query=workflow%3ACI)
-[![Coverage Status](https://coveralls.io/repos/JuliaDiffEq/SciMLExpectations.jl/badge.svg?branch=master&service=github)](https://coveralls.io/github/JuliaDiffEq/SciMLExpectations.jl?branch=master)
-[![codecov.io](http://codecov.io/github/JuliaDiffEq/SciMLExpectations.jl/coverage.svg?branch=master)](http://codecov.io/github/JuliaDiffEq/SciMLExpectations.jl?branch=master)
+[![Coverage Status](https://coveralls.io/repos/SciML/SciMLExpectations.jl/badge.svg?branch=master&service=github)](https://coveralls.io/github/SciML/SciMLExpectations.jl?branch=master)
+[![codecov.io](http://codecov.io/github/SciML/SciMLExpectations.jl/coverage.svg?branch=master)](http://codecov.io/github/JuliaDiffEq/SciMLExpectations.jl?branch=master)
 
 SciMLExpectations.jl is a package for quantifying the uncertainties of simulations by
 calculating the expectations of observables with respect to input uncertainties. Its goal
