+++ src/SciMLExpectations.jl
@@ -4,15 +4,16 @@ module SciMLExpectations
 using DiffEqBase, Statistics, Reexport, RecursiveArrayTools, StaticArrays,
       Distributions, KernelDensity, Zygote, LinearAlgebra, Random
 using Parameters: @unpack
+import DiffEqNoiseProcess
 
 @reexport using Integrals
 import DiffEqBase: solve
 
-include("system_utils.jl")
 include("distribution_utils.jl")
 include("problem_types.jl")
 include("solution_types.jl")
 include("expectation.jl")
+include("system_utils.jl")
 
 # Type Piracy, should upstream
 Base.eltype(K::UnivariateKDE) = eltype(K.density)
@@ -29,6 +30,7 @@ Base.maximum(d::Product) = maximum.(d.v)
 Base.extrema(d::Product) = minimum(d), maximum(d)
 
 export Koopman, MonteCarlo, PrefusedAD, PostfusedAD, NonfusedAD
-export GenericDistribution, SystemMap, ExpectationProblem, build_integrand
+export GenericDistribution, SystemMap, ProcessNoiseSystemMap, ExpectationProblem,
+       build_integrand
 
 end
