+++ src/problem_types.jl
@@ -38,13 +38,6 @@ function ExpectationProblem(g, pdist, params; nout = 1)
     ExpectationProblem(S, g, h, pdist, params, nout)
 end
 
-# Constructor for DEProblems
-function ExpectationProblem(sm::SystemMap, g, h, d; nout = 1)
-    ExpectationProblem(sm, g, h, d,
-                       ArrayPartition(deepcopy(sm.prob.u0), deepcopy(sm.prob.p)),
-                       nout)
-end
-
 distribution(prob::ExpectationProblem) = prob.d
 mapping(prob::ExpectationProblem) = prob.S
 observable(prob::ExpectationProblem) = prob.g
